COMP4442/project/requirements.txt
flask==2.2.5
flask-wtf==1.0.1
boto3==1.28.39
python-dotenv==1.0.0
requests==2.31.0
Werkzeug==2.2.3
flask-login==0.6.2
email_validator==2.0.0
flask-session==0.4.0
markupsafe==2.1.3
itsdangerous==2.1.2
Jinja2==3.1.2
gunicorn==20.1.0
